+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Ability to merge registries
 
 ### Changed
+- Correct typedefs for object constructor of metrics
 
 ## [9.0.0] - 2017-05-06
 ### Added